Expose a py_image rule for packaging a py_binary target as a Docker image.
This builds on the mattmoor/rules_docker foo-image branch to create a specialized form of py_binary that packages the resulting application (atop the distroless python runtime) as a Docker image.

Each of the dependencies' runfiles are split into its own layer and the filesystem layout is designed such that these layers may be shared between distinct apps with the same shared dependency.

At this point, this is purely a proof of concept, and there are a fair number of TODOs sprinkled throughout.